tree: Extract tree functions
To support blaming trees, extract show_tree_common and show_recursive
into tree.h.
